增加基类
